@PkmX 2025-03-29
I'm getting this error when running https://huggingface.co/ggml-org/gemma-3-4b-it-GGUF which `git blame` points to this PR:

```
ggml_metal_init: allocating
ggml_metal_init: found device: Apple M1 Ultra
ggml_metal_init: picking default device: Apple M1 Ultra
ggml_metal_load_library: using embedded metal library
ggml_metal_load_library: error: Error Domain=MTLLibraryErrorDomain Code=3 "program_source:6106:12: error: variable length arrays are not supported in Metal
    o4_t lo[DV4/NL];
           ^
program_source:6405:18: note: in instantiation of function template specialization 'kernel_flash_attn_ext_vec<half __attribute__((ext_vector_type(4))), half __attribute__((ext_vector_type(4))), half __attribute__((ext_vector_type(4))), float, half, half __attribute__((ext_vector_type(4))), half __attribute__((ext_vector_type(4))), half __attribute__((ext_vector_type(4))), 1, &dequantize_f16_t4, half __attribute__((ext_vector_type(4))), 1, &dequantize_f16_t4, 128, 128, 128, 1, 32>' requested here
typedef decltype(kernel_flash_attn_ext_vec<FA_TYPES, half4, 1, dequantize_f16_t4, half4, 1, dequantize_f16_t4, 128, 128, 128>) flash_attn_ext_vec_t;
                 ^
" UserInfo={NSLocalizedDescription=program_source:6106:12: error: variable length arrays are not supported in Metal
    o4_t lo[DV4/NL];
           ^
program_source:6405:18: note: in instantiation of function template specialization 'kernel_flash_attn_ext_vec<half __attribute__((ext_vector_type(4))), half __attribute__((ext_vector_type(4))), half __attribute__((ext_vector_type(4))), float, half, half __attribute__((ext_vector_type(4))), half __attribute__((ext_vector_type(4))), half __attribute__((ext_vector_type(4))), 1, &dequantize_f16_t4, half __attribute__((ext_vector_type(4))), 1, &dequantize_f16_t4, 128, 128, 128, 1, 32>' requested here
typedef decltype(kernel_flash_attn_ext_vec<FA_TYPES, half4, 1, dequantize_f16_t4, half4, 1, dequantize_f16_t4, 128, 128, 128>) flash_attn_ext_vec_t;
                 ^
}
ggml_metal_init: error: metal library is nil
ggml_backend_metal_device_init: error: failed to allocate context
llama_init_from_model: failed to initialize the context: failed to initialize Metal backend
common_init_from_params: failed to create context with model 'models/gemma-3-4b-it-Q8_0.gguf'
srv    load_model: failed to load model, 'models/gemma-3-4b-it-Q8_0.gguf'
srv    operator(): operator(): cleaning up before exit...
libc++abi: terminating
main: exiting due to model loading error
```