@ggerganov 2025-03-30
@PkmX Can you confirm that https://github.com/ggml-org/llama.cpp/pull/12659 fixes the error?