@PkmX 2025-03-30
@ggerganov No, still getting the same error even after a clean rebuild.