@ggerganov 2025-03-30
@PkmX I pushed one more change there - can you check if it works now?

It's strange why it fails for you. I tested that the `master` works correctly on M1 Pro, M2 Ultra and M4 Pro.